Speed up time to load
We were waiting on the DOM to finish loading, which meant all images
and assets had to load before the link would appear. On my (not very good)
internet connection, you'd have to wait quite a while before you could
jump to the link.

Because we've set the content script to run at `document_end`, we
actually don't need to wait for the DOM to load. `document_end` means
the script is “injected immediately after the DOM is complete, but
before subresources like images and frames have loaded”.

This change removes the DOM wait, and removes the jQuery injection while
its at it.